Disable direct reference to outer scope field in SELECT
Before this change, a query with direct reference to outer scope field
in `SELECT` clause failed with `given correlated subquery not supported`.
Likewise, any non-direct field reference (projection) on correlated
symbol failed. In both cases, the failure was caused by `ProjectNode`'s
assignments not being decorrelated.
The goal of this change is to enable decorrelating projections, except for
the special case of direct correlated field reference, which cannot be
handled correctly in the current implementation of correlation.