Refactor UnaliasSymbolReferences
introduce context, including:
    - inherited mapping for correlation symbols
    - symbols not allowed in given context ("forbidden symbols")
      (if encountered, will be remapped to a new symbol)
introduce return type for visit() methods, including:
    - rewritten plan
    - symbol mappings (both inherited and derived)

Also introduce new SymbolMapper, AliasingSymbolMapper
In this mapper, mapping can be modified when "forbidden symbols" are defined.
Whenever a yet unmapped "forbidden symbol" is encountered, it is mapped
to a new symbol, and this mapping is added to mappings.
In case when "forbidden symbols" are not specified, AliasingSymbolMapper
is umnodifiable and behaves the same way as SymbolMapper. However, it
supports mapping for more node kinds, and is cleaner.

TODO
- remove "forbidden symbols" handling
- replace SymbolMapper with the new mapper